Split lib.rs into many files
